Fixed pixel look up, but still throws GL errors on debug builds.

+++ mrv2/lib/mrvGL/mrvTimelineWidget.h
@@ -138,7 +138,7 @@ namespace mrv
         void _drawAnnotationMarks() const noexcept;
 
         otime::RationalTime _posToTime(int) const noexcept;
-        double _timeToPos(const otime::RationalTime&) const noexcept;
+        int _timeToPos(const otime::RationalTime&) const noexcept;
 
         void _styleUpdate();
